Move the CLI application into its own repo; flattens workspace
