--- .vscode/launch.json
@@ -1,19 +1,0 @@
-{
-  // Use IntelliSense to learn about possible attributes.
-  // Hover to view descriptions of existing attributes.
-  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "name": "debug build",
-      "type": "cppvsdbg",
-      "request": "launch",
-      "program": "${workspaceFolder}/target/debug/etherdream-cli.exe",
-      "args": [ "build" ],
-      "stopAtEntry": false,
-      "cwd": "${workspaceFolder}",
-      "environment": [],
-      "console": "integratedTerminal"
-    }
-  ]
-}
