--- etherdream-cli/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-edition = "2024"
-license = "MIT"
-name = "etherdream-cli"
-version = "0.1.0"
-
-[dependencies]
-clap = { version = "4.5.35", features = ["derive"] }
-etherdream = { path = "../etherdream" }
-shlex = { version="1.3.0" }
-tokio = { workspace=true, features=["full"] }
