--- etherdream-test/src/lib.rs
@@ -1,5 +1,0 @@
-pub mod etherdream_emulator;
-pub mod test_device;
-
-pub use etherdream_emulator::*;
-pub use test_device::*;
