--- etherdream/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-edition = "2024"
-license = "MIT"
-name = "etherdream"
-version = "0.1.0"
-
-[dependencies]
-tokio = { workspace = true }
-tokio-util = { version = "0.7.16", features = ["net"] }
-
-[dev-dependencies]
-etherdream-test = { path = "../etherdream-test" }
-tokio = { workspace = true, features = ["full"] }
