+++ src/lib.rs
@@ -5,6 +5,6 @@ pub mod device;
 pub mod discovery;
 
 // Convenience exports
-pub use client::{ Client, ClientBuilder, Point };
+pub use client::{Client, ClientBuilder, Point};
 pub use device::Device;
-pub use discovery::{ Server as Discovery };
+pub use discovery::Server as Discovery;
